띄워지는종류들 다 컴포넌트로 구분함함
