+++ src/router/Router.tsx
@@ -12,6 +12,7 @@ import MypageSettingManage from '../pages/mypage/mypage-Setting/MypageSettingMan
 import MypageFeedContent from '../pages/mypage/mypage-main/mypage-feed/MypageFeedContent';
 import MypageRecord from '../pages/mypage/mypage-record/MypageRecord';
 import BottomNavi from '../components/BottomNavi/BottomNavi';
+import CustomMain from '../pages/custom/custom';
 
 const Router = () => {
   const router = createBrowserRouter([
@@ -52,11 +53,16 @@ const Router = () => {
           element: <BottomNavi />,
         },
         
+        
         // {BottomNavi
         //   path: '/edit/completed',
         //   element: <EditCompleted />
         // }
       ],
+    },{
+      path: '/custom',
+      element: <CustomMain brand="스타벅스" drink="아이스 아메리카노" sugar={0} kcal={0} caffeine={5}/>,
+
     },
   ]);
